Merge release (2011a1905) into master: [SV4-85] Allow zero observations during RTCM -> SBP conversion (#756) (#767)
Version 5.8.14 of the TeseoV firmware introduced a configuration option which allows MSM7 messages to be sent when no satellites are available. Enabling this setting allows the possibility of outputting MSG_GPS_TIME_GNSS messages from the Starling GNSS engine in situations where no satellites are in view (or the antenna is disconnected). This is required in order to allow synchronisation of external sensor data with Starling time during GNSS signal blockage events

Co-authored-by: Jason Mobarak <jason@swiftnav.com>

* Removing submodule updates

Co-authored-by: Jason Mobarak <jason@swiftnav.com>
Co-authored-by: Rodrigo Reichert <rodrigo.reichert@swift-nav.com>
Co-authored-by: dgburr <dburr@dburr.net>